refactor(rules): split isOwner.ts god file into one rule per file
isOwner.ts had grown to ~700 lines holding 10 independent graphql-shield
ownership/authorship rules for 12+ entity types. Move each rule into its own
file under rules/permission/ownership/ and turn isOwner.ts into a barrel that
re-exports them, so the ./permission/isOwner.js import surface (used by
rules/rules.ts) is unchanged.

Pure mechanical extraction — no logic changes. tsc clean, 565/565 unit tests
pass.
