+++ rules/permission/isOwner.ts
@@ -1,719 +1,15 @@
-import { rule } from "graphql-shield";
-import type { GraphQLResolveInfo } from "graphql";
-import type { GraphQLContext } from "../../types/context.js";
-import { ERROR_MESSAGES } from "../errorMessages.js";
-import {
-  ChannelWhere,
-  Channel,
-  Comment,
-  Discussion,
-  Event,
-  DiscussionWhere,
-  DiscussionChannelWhere,
-  DiscussionUpdateInput,
-  EventWhere,
-  EventUpdateInput,
-  CommentWhere,
-  UserWhere,
-  User,
-  Collection,
-  CollectionWhere,
-  Image,
-  ImageWhere,
-  Issue,
-  IssueWhere,
-} from "../../src/generated/graphql.js";
-import { setUserDataOnContext } from "./userDataHelperFunctions.js";
-import { logger } from "../../logger.js";
-
-type IsChannelOwnerInput = {
-  where: ChannelWhere;
-  channelUniqueName: string;
-  issueId?: string;
-  commentId?: string;
-};
-
-export const isChannelOwner = rule({ cache: "contextual" })(
-  async (parent: unknown, args: IsChannelOwnerInput, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    logger.info('🔐 isChannelOwner rule called with args:', JSON.stringify(args));
-
-    // set user data
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-    let username = ctx.user.username;
-    logger.info("username: ", ctx.user);
-
-    let ogm = ctx.ogm;
-    const { where, channelUniqueName, issueId, commentId } = args;
-    logger.info('args: ', JSON.stringify(args));
-    let uniqueName = '';
-
-    if (where?.uniqueName) {
-      // The channel name can be passed in the where object.
-      uniqueName = where.uniqueName;
-    }
-
-    if (channelUniqueName) {
-      // It can also be passed as a separate argument.
-      uniqueName = channelUniqueName;
-    }
-
-    // If no channel name is provided but we have a commentId, look it up
-    if (!uniqueName && commentId) {
-      const Comment = ogm.model("Comment");
-      const comment = await Comment.find({
-        where: { id: commentId },
-        selectionSet: `{ 
-          DiscussionChannel {
-            channelUniqueName
-          }
-          Channel {
-            uniqueName
-          }
-        }`,
-      });
-
-      if (!comment || !comment[0]) {
-        throw new Error(ERROR_MESSAGES.channel.notFound);
-      }
-
-      // Try to get channel name from either DiscussionChannel or Channel
-      uniqueName = comment[0]?.DiscussionChannel?.channelUniqueName ||
-                  comment[0]?.Channel?.uniqueName || "";
-
-      if (!uniqueName) {
-        throw new Error(ERROR_MESSAGES.channel.notFound);
-      }
-    }
-
-    // If no channel name is provided but we have an issueId, look it up
-    // Also check where.id for mutations like updateIssues that pass issue ID in where clause
-    const resolvedIssueId = issueId || (where as any)?.id;
-    if (!uniqueName && resolvedIssueId) {
-      const Issue = ogm.model("Issue");
-      const issue = await Issue.find({
-        where: { id: resolvedIssueId },
-        selectionSet: `{
-          channelUniqueName
-        }`,
-      });
-      logger.info('issue', JSON.stringify(issue));
-
-      if (!issue || !issue[0]) {
-        // Return false instead of throwing to allow OR chain to continue
-        return false;
-      }
-
-      uniqueName = issue[0].channelUniqueName ?? "";
-    }
-
-    if (!uniqueName) {
-      // Return false instead of throwing to allow OR chain to continue
-      return false;
-    }
-
-    const ChannelModel = ogm.model("Channel");
-
-    // Get the list of channel owners by using the OGM on the
-    // Channel object.
-    const channel: Channel[] = await ChannelModel.find({
-      where: { uniqueName },
-      selectionSet: `{ 
-            Admins { 
-                username
-            } 
-      }`,
-    });
-
-    if (!channel) {
-      throw new Error(ERROR_MESSAGES.channel.notFound);
-    }
-
-    if (channel.length === 0) {
-      throw new Error(ERROR_MESSAGES.channel.notFound);
-    }
-
-    // Get the list of channel owners.
-    const channelOwners = channel[0].Admins.map((admin) => admin.username);
-
-    // Check if the user is in the list of channel owners.
-    if (!channelOwners.includes(username ?? "")) {
-      return false;  // Permission check - return false to allow OR to work
-    }
-
-    return true;
-  }
-);
-
-type IsDiscussionOwnerInput = {
-  where: DiscussionWhere;
-  discussionUpdateInput: DiscussionUpdateInput;
-  channelConnections: string[];
-  channelDisconnections: string[];
-};
-
-export const isDiscussionOwner = rule({ cache: "contextual" })(
-  async (parent: unknown, args: IsDiscussionOwnerInput, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    let discussionId;
-
-    const { where } = args;
-    if (where) {
-      discussionId = where.id;
-    }
-
-    // set user data
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-
-    let username = ctx.user.username;
-    let ogm = ctx.ogm;
-
-    if (!discussionId) {
-      throw new Error(ERROR_MESSAGES.discussion.noId);
-    }
-    const DiscussionModel = ogm.model("Discussion");
-
-    // Get the discussion owner by using the OGM on the
-    // Discussion model.
-    const discussions: Discussion[] = await DiscussionModel.find({
-      where: { id: discussionId },
-      selectionSet: `{ 
-            Author { 
-                username
-            } 
-      }`,
-    });
-
-    if (!discussions || discussions.length === 0) {
-      throw new Error(ERROR_MESSAGES.channel.notFound);
-    }
-    const discussion = discussions[0];
-
-    // Get the discussion author.
-    const discussionOwner = discussion?.Author?.username;
-
-    if (!discussionOwner) {
-      throw new Error(ERROR_MESSAGES.discussion.noAuthor);
-    }
-
-    // Check if the user is the discussion owner
-    if (discussionOwner !== username) {
-      return false;  // Permission check - return false to allow OR to work
-    }
-    return true;
-  }
-);
-
-type IsEventOwnerInput = {
-  where: EventWhere;
-  eventUpdateInput: EventUpdateInput;
-  channelConnections: string[];
-  channelDisconnections: string[];
-};
-
-export const isEventOwner = rule({ cache: "contextual" })(
-  async (parent: unknown, args: IsEventOwnerInput, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-
-    let eventId;
-
-    const { where } = args;
-    if (where) {
-      eventId = where.id;
-    }
-
-    // set user data
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-
-    let username = ctx.user.username;
-    let ogm = ctx.ogm;
-
-    if (!eventId) {
-      throw new Error(ERROR_MESSAGES.event.noId);
-    }
-    const EventModel = ogm.model("Event");
-
-    // Get the event owner by using the OGM on the
-    // Event model.
-    const events: Event[] = await EventModel.find({
-      where: { id: eventId },
-      selectionSet: `{ 
-            Poster { 
-                username
-            } 
-      }`,
-    });
-
-    if (!events || events.length === 0) {
-      throw new Error(ERROR_MESSAGES.event.notFound);
-    }
-    const event = events[0];
-
-    // Get the event author.
-    const eventOwner = event?.Poster?.username;
-
-    if (!eventOwner) {
-      throw new Error(ERROR_MESSAGES.event.noOwner);
-    }
-
-    // Check if the user is in the list of channel owners.
-    if (eventOwner !== username) {
-      return false;  // Permission check - return false to allow OR to work
-    }
-    return true;
-  }
-);
-
-type IsCommentAuthorInput = {
-  where: CommentWhere;
-  update: CommentWhere;
-};
-
-export const isCommentAuthor = rule({ cache: "contextual" })(
-  async (parent: unknown, args: IsCommentAuthorInput, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    logger.info("isCommentAuthor rule");
-    const { where } = args;
-    const commentId  = where.id
-
-    // set user data
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-
-    let username = ctx.user.username;
-    let modName =  ctx.user.data?.ModerationProfile?.displayName || null;
-    logger.info("username: ", username);
-    logger.info("modName: ", modName);
-
-    let ogm = ctx.ogm;
-
-    if (!commentId) {
-      throw new Error(ERROR_MESSAGES.comment.noId);
-    }
-    const CommentModel = ogm.model("Comment");
-
-    // Get the comment owner by using the OGM on the
-    // Comment model.
-    const comments: Comment[] = await CommentModel.find({
-      where: { id: commentId },
-      selectionSet: `{ 
-        CommentAuthor {
-          ... on User {
-            username
-          }
-          ... on ModerationProfile {
-            displayName
-          }
-        }
-      }`,
-    });
-
-    if (!comments || comments.length === 0) {
-      throw new Error(ERROR_MESSAGES.comment.notFound);
-    }
-    const comment = comments[0];
-
-    // Get the comment author.
-    const author = comment?.CommentAuthor;
-    let authorUsername;
-    let authorModProfileName;
-
-    // The comment owner could be a user or a moderation profile.
-    // For a user, the username is stored on the user object.
-    // For a moderation profile, the displayName is stored on 
-    // the moderation profile object.
-    if (!author) {
-      throw new Error(ERROR_MESSAGES.comment.noOwner);
-    }
-
-    // @ts-ignore
-    if (author.username) {
-      // @ts-ignore
-      authorUsername = author.username;
-    } else if (author.displayName) {
-      authorModProfileName = author.displayName;
-    } else {
-      throw new Error(ERROR_MESSAGES.comment.noOwner);
-    }
-
-    // Check if the user is the comment author.
-    if (authorUsername && authorUsername !== username) {
-      return false;  // Permission check - return false to allow OR to work
-    }
-    if (authorModProfileName && authorModProfileName !== modName) {
-      return false;  // Permission check - return false to allow OR to work
-    }
-    return true;
-  }
-);
-
-interface IsAccountOwnerArgs {
-  where: UserWhere;
-  username: string;
-}
-// Check if the user is the owner of the account.
-export const isAccountOwner = rule({ cache: "contextual" })(
-  async (parent: { username?: string } | undefined, args: IsAccountOwnerArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    // When accessing nested fields like FavoriteChannels, the username comes from the parent User object
-    const username  = parent?.username || args.where?.username || args.username;
-
-    // set user data
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-
-    if (!username) {
-      throw new Error(ERROR_MESSAGES.user.noUsername);
-    }
-
-    // Check if the user is the account owner.
-    if (username !== ctx.user.username) {
-      throw new Error(ERROR_MESSAGES.user.notOwner);
-    }
-
-    return true;
-  }
-);
-
-type IsCollectionOwnerArgs = {
-  where?: CollectionWhere;
-  collectionId?: string;
-  id?: string;
-};
-
-export const isCollectionOwner = rule({ cache: "contextual" })(
-  async (parent: { id?: string } | undefined, args: IsCollectionOwnerArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-
-    const username = ctx.user?.username;
-
-    if (!username) {
-      throw new Error(ERROR_MESSAGES.user.noUsername);
-    }
-
-    const collectionIds: string[] = [];
-    const whereArg = args?.where;
-
-    if (whereArg?.id) {
-      collectionIds.push(whereArg.id);
-    }
-
-    if (whereArg?.id_IN && Array.isArray(whereArg.id_IN)) {
-      collectionIds.push(...whereArg.id_IN);
-    }
-
-    if (args?.collectionId) {
-      collectionIds.push(args.collectionId);
-    }
-
-    if (args?.id) {
-      collectionIds.push(args.id);
-    }
-
-    if (parent?.id && collectionIds.length === 0) {
-      collectionIds.push(parent.id);
-    }
-
-    if (collectionIds.length === 0) {
-      throw new Error(ERROR_MESSAGES.collection.noId);
-    }
-
-    const uniqueIds = [...new Set(collectionIds)];
-
-    const CollectionModel = ctx.ogm.model("Collection");
-    const whereClause: CollectionWhere =
-      uniqueIds.length === 1
-        ? { id: uniqueIds[0] }
-        : { id_IN: uniqueIds };
-
-    const collections: Collection[] = await CollectionModel.find({
-      where: whereClause,
-      selectionSet: `{ id CreatedBy { username } }`,
-    });
-
-    if (!collections || collections.length === 0) {
-      throw new Error(ERROR_MESSAGES.collection.notFound);
-    }
-
-    if (collections.length !== uniqueIds.length) {
-      throw new Error(ERROR_MESSAGES.collection.notFound);
-    }
-
-    const isOwner = collections.every(
-      (collection) => collection?.CreatedBy?.username === username
-    );
-
-    if (!isOwner) {
-      throw new Error(ERROR_MESSAGES.collection.notOwner);
-    }
-
-    return true;
-  }
-);
-
-type IsDiscussionChannelOwnerInput = {
-  where: DiscussionChannelWhere;
-};
-
-export const isDiscussionChannelOwner = rule({ cache: "contextual" })(
-  async (parent: unknown, args: IsDiscussionChannelOwnerInput, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    // Set user data
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-
-    let username = ctx.user.username;
-    let ogm = ctx.ogm;
-
-    // Extract the discussionId and channelUniqueName from the various possible argument formats
-    let discussionId;
-    let channelUniqueName;
-
-    if (args.where) {
-      discussionId = args.where.discussionId;
-      channelUniqueName = args.where.channelUniqueName;
-    }
-
-    if (!discussionId || !channelUniqueName) {
-      logger.error("Missing discussionId or channelUniqueName in args:", args);
-      return false;
-    }
-
-    try {
-      // First check if the user is the owner of the discussion
-      const DiscussionModel = ogm.model("Discussion");
-      const discussions = await DiscussionModel.find({
-        where: { id: discussionId },
-        selectionSet: `{
-          Author {
-            username
-          }
-        }`,
-      });
-
-      const discussionAuthor = discussions[0]?.Author?.username
-
-      if (discussions && discussions.length > 0) {
-        const discussionOwner = discussionAuthor;
-        if (discussionOwner === username) {
-          return true;
-        }
-      }
-
-      // Next check if the user is an admin of the channel
-      const ChannelModel = ogm.model("Channel");
-      const channels = await ChannelModel.find({
-        where: { uniqueName: channelUniqueName },
-        selectionSet: `{
-          Admins {
-            username
-          }
-        }`,
-      });
-
-      if (channels && channels.length > 0) {
-        const channelAdmins = channels[0].Admins.map((admin: User) => admin.username);
-        if (channelAdmins.includes(username ?? "")) {
-          return true;
-        }
-      }
-
-      return false;
-    } catch (error) {
-      logger.error("Error in isDiscussionChannelOwner:", error);
-      return false;
-    }
-  }
-);
-
-type IsImageUploaderInput = {
-  where: ImageWhere;
-};
-
-export const isImageUploader = rule({ cache: "contextual" })(
-  async (parent: unknown, args: IsImageUploaderInput, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const { where } = args;
-    const imageId = where?.id;
-
-    // Set user data
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-
-    const username = ctx.user.username;
-    const ogm = ctx.ogm;
-
-    if (!imageId) {
-      throw new Error(ERROR_MESSAGES.image.noId);
-    }
-
-    const ImageModel = ogm.model("Image");
-
-    // Get the image and its uploader
-    const images: Image[] = await ImageModel.find({
-      where: { id: imageId },
-      selectionSet: `{
-        id
-        Uploader {
-          username
-        }
-      }`,
-    });
-
-    logger.info('🔍 isImageUploader - Found images:', JSON.stringify(images, null, 2));
-
-    if (!images || images.length === 0) {
-      throw new Error(ERROR_MESSAGES.image.notFound);
-    }
-
-    const image = images[0];
-    const uploaderUsername = image?.Uploader?.username;
-
-    logger.info('🔍 isImageUploader - Image:', { imageId: image.id, uploaderUsername, loggedInUsername: username });
-
-    // If there's no uploader set (legacy/test data), allow anyone authenticated to edit
-    // This maintains backward compatibility with existing images
-    if (!uploaderUsername) {
-      logger.info('⚠️ isImageUploader - No uploader found, allowing edit (legacy data)');
-      return true;
-    }
-
-    // Check if the user is the uploader
-    if (uploaderUsername !== username) {
-      logger.info('❌ isImageUploader - User is not the uploader');
-      return false; // Permission check - return false to allow OR to work
-    }
-
-    logger.info('✅ isImageUploader - User is the uploader');
-    return true;
-  }
-);
-
-type IsIssueAuthorInput = {
-  where: IssueWhere;
-};
-
-/**
- * Check if the current user is the author of the issue.
- * The issue author can be either a User or a ModerationProfile.
- */
-export const isIssueAuthor = rule({ cache: "contextual" })(
-  async (parent: unknown, args: IsIssueAuthorInput, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const { where } = args;
-    const issueId = where?.id;
-
-    // Set user data
-    ctx.user = await setUserDataOnContext({
-      context: ctx,
-      getPermissionInfo: false,
-    });
-
-    const username = ctx.user.username;
-    const modName = ctx.user.data?.ModerationProfile?.displayName || null;
-    const ogm = ctx.ogm;
-
-    if (!issueId) {
-      throw new Error(ERROR_MESSAGES.issue.noId);
-    }
-
-    const IssueModel = ogm.model("Issue");
-
-    // Get the issue author using the OGM
-    const issues: Issue[] = await IssueModel.find({
-      where: { id: issueId },
-      selectionSet: `{
-        Author {
-          ... on User {
-            username
-          }
-          ... on ModerationProfile {
-            displayName
-          }
-        }
-      }`,
-    });
-
-    if (!issues || issues.length === 0) {
-      throw new Error(ERROR_MESSAGES.issue.notFound);
-    }
-
-    const issue = issues[0];
-    const author = issue?.Author;
-
-    if (!author) {
-      throw new Error(ERROR_MESSAGES.issue.noAuthor);
-    }
-
-    // The issue author could be a user or a moderation profile
-    // @ts-ignore - Author union type
-    const authorUsername = author.username;
-    // @ts-ignore - Author union type
-    const authorModProfileName = author.displayName;
-
-    // Check if the current user matches the author
-    if (authorUsername && authorUsername === username) {
-      return true;
-    }
-    if (authorModProfileName && authorModProfileName === modName) {
-      return true;
-    }
-
-    return false; // Permission check - return false to allow OR to work
-  }
-);
-
-type IssueIsNotLockedInput = {
-  where: IssueWhere;
-};
-
-// Extended Issue type to include locked field (exists in schema but not yet in generated types)
-type IssueWithLocked = Issue & { locked?: boolean };
-
-/**
- * Check if the issue is not locked.
- * Locked issues cannot be modified by OPs (only by moderators).
- */
-export const issueIsNotLocked = rule({ cache: "contextual" })(
-  async (parent: unknown, args: IssueIsNotLockedInput, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const { where } = args;
-    const issueId = where?.id;
-    const ogm = ctx.ogm;
-
-    if (!issueId) {
-      throw new Error(ERROR_MESSAGES.issue.noId);
-    }
-
-    const IssueModel = ogm.model("Issue");
-
-    const issues = (await IssueModel.find({
-      where: { id: issueId },
-      selectionSet: `{ locked }`,
-    })) as unknown as IssueWithLocked[];
-
-    if (!issues || issues.length === 0) {
-      throw new Error(ERROR_MESSAGES.issue.notFound);
-    }
-
-    const issue = issues[0];
-
-    // If locked, deny the action (return false to allow OR with mod permissions)
-    if (issue.locked) {
-      return false;
-    }
-
-    return true;
-  }
-);
+// Barrel for the ownership/authorship graphql-shield rules.
+//
+// Each rule previously lived in this file (it had grown to ~700 lines covering
+// 12+ entity types). They now live one-per-file under ./ownership/; this barrel
+// preserves the original `./permission/isOwner.js` import surface.
+export { isChannelOwner } from "./ownership/isChannelOwner.js";
+export { isDiscussionOwner } from "./ownership/isDiscussionOwner.js";
+export { isEventOwner } from "./ownership/isEventOwner.js";
+export { isCommentAuthor } from "./ownership/isCommentAuthor.js";
+export { isAccountOwner } from "./ownership/isAccountOwner.js";
+export { isCollectionOwner } from "./ownership/isCollectionOwner.js";
+export { isDiscussionChannelOwner } from "./ownership/isDiscussionChannelOwner.js";
+export { isImageUploader } from "./ownership/isImageUploader.js";
+export { isIssueAuthor } from "./ownership/isIssueAuthor.js";
+export { issueIsNotLocked } from "./ownership/issueIsNotLocked.js";
